tests: harness: move filter.h to the main test harness in test.h
