+++ tests/unit/CMakeLists.txt
@@ -90,7 +90,6 @@ add_executable(unit_tests
     assert_override.h
 
     harness/main.c
-    harness/filter.c harness/filter.h
     harness/helper.c harness/helper.h
     harness/opts.c harness/opts.h
