+++ tests/unit/harness/opts.h
@@ -5,13 +5,15 @@
 
 #pragma once
 
+#include "harness/test.h"
+
 #define _cleanup_bf_test_opts_ __attribute__((cleanup(bf_test_opts_free)))
 
 struct bf_test_filter;
 
 struct bf_test_opts
 {
-    struct bf_test_filter *group_filter;
+    bf_test_filter *group_filter;
 };
 
 int bf_test_opts_new(struct bf_test_opts **opts, int argc, char *argv[]);
